requirements.txt
fastapi
uvicorn[standard]
streamlit
plotly
pandas
numpy
python-dotenv
requests
